# src/main/resources/swagger.yaml
openapi: 3.0.2
info:
  title: Microservicio de Products
  version: v1
paths:
  '/stores/{storeId}':
    parameters:
      - $ref: '#/components/parameters/storeId'
    delete:
      summary: Delete a store by id.
      operationId: deleteStoreById
      responses:
        '204':
          description: Store created successfully, no content returned.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    get:
      summary: Get a store by id.
      operationId: getStoreById
      responses:
        '200':
          $ref: '#/components/responses/Store'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /stores:
    get:
      summary: Get all stores
      operationId: getStores
      responses:
        '200':
          $ref: '#/components/responses/ListStores'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Create a Store
      operationId: createStore
      requestBody:
        $ref: '#/components/requestBodies/StoreRequestDTO'
      responses:
        '201':
          description: Store created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /products:
    get:
      summary: Get a list of products by StoreId
      operationId: getProductsByStoreId
      responses:
        '200':
          $ref: '#/components/responses/ListProducts'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Create a product to current Store
      operationId: createProduct
      requestBody:
        $ref: '#/components/requestBodies/ProductRequestDTO'
      responses:
        '201':
          description: Store created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /categories:
    get:
      summary: Get a list of categories by StoreId
      operationId: getCategoriesByStoreId
      responses:
        '200':
          $ref: '#/components/responses/ListCategories'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Create a category
      operationId: createCategory
      requestBody:
        $ref: '#/components/requestBodies/CategoryRequestDTO'
      responses:
        '201':
          description: Category created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
components:
  requestBodies:
    StoreRequestDTO:
      description: Create a new store
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StoreRequestDTO'
    ProductRequestDTO:
      description: Create a new product
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductRequestDTO'
    CategoryRequestDTO:
      description: Create a new category
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CategoryRequestDTO'
  parameters:
    productId:
      name: productId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/productId'
    categoryId:
      name: categoryId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/categoryId'
    storeId:
      name: storeId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/storeId'
  responses:
    ListStores:
      description: OK.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/StoreDTO'
    Store:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StoreDTO'
    ListProducts:
      description: OK.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ProductDTO'
    ListCategories:
      description: OK.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CategoryDTO'
    Product:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductDTO'
    Category:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CategoryDTO'
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 400
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Unauthorized:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 401
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Forbidden:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 403
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    NotFound:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    UnexpectedError:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
  schemas:
    CategoryRequestDTO:
      type: object
      properties:
        categoryName:
          type: string
          example: "Drinks"
        storeId:
          type: integer
          format: int64
    ProductRequestDTO:
      type: object
      properties:
        productName:
          type: string
          example: "Cafe Latte"
        price:
          type: string
          format: number
          example: "4.99"
        categoryId:
          type: integer
          format: int64
          example: 12345
        storeId:
          type: integer
          format: int64
    StoreRequestDTO:
      type: object
      properties:
        name:
          type: string
          example: New York Store
        address:
          $ref: '#/components/schemas/AddressDTO'
    StoreDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        name:
          type: string
          example: New York Store
        address:
          $ref: '#/components/schemas/AddressDTO'
        createdAt:
          type: string
          format: date-time
          example: "2023-10-01T12:34:56Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-01T12:34:56Z"
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        categoryName:
          type: string
          example: "Drinks"
        storeId:
          type: integer
          format: int64
          example: 12345
        createdAt:
          type: string
          format: date-time
          example: "2023-10-01T12:34:56Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-01T12:34:56Z"
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        productName:
          type: string
          example: "Cafe Latte"
        price:
          type: string
          format: number
          example: "4.99"
        categoryId:
          type: integer
          format: int64
          example: 12345
        storeId:
          type: integer
          format: int64
          example: 12345
        createdAt:
          type: string
          format: date-time
          example: "2023-10-01T12:34:56Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-10-01T12:34:56Z"
    AddressDTO:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        country:
          type: string
          example: "USA"
    Problem:
      type: object
      properties:
        error:
          type: string
    productId:
      title: productId
      description: productId
      type: integer
      format: int64
      minimum: 1
      maximum: 999999999999
      example: 1
    categoryId:
      title: example
      description: example
      type: integer
      format: int64
      minimum: 1
      maximum: 999999999999
      example: 1
    storeId:
      title: storeId
      description: storeId
      type: integer
      format: int64
      minimum: 1
      maximum: 999999999999
      example: 1
    saleId:
      title: sateId
      description: sateId
      type: integer
      format: int64
      minimum: 1
      maximum: 999999999999
      example: 1